renamed (:feature:now) to "tomorrow", fixed related bugs. Cleared all navigation graphs
