+++ .idea/gradle.xml
@@ -17,8 +17,8 @@
             <option value="$PROJECT_DIR$/core/ui" />
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/init-location" />
-            <option value="$PROJECT_DIR$/feature/now" />
             <option value="$PROJECT_DIR$/feature/today" />
+            <option value="$PROJECT_DIR$/feature/tomorrow" />
             <option value="$PROJECT_DIR$/feature/week" />
             <option value="$PROJECT_DIR$/weather_api" />
           </set>
