--- app/src/main/java/ru/msokolov/weatherapp/MainFragment.kt
@@ -1,25 +1,0 @@
-package ru.msokolov.weatherapp
-
-import android.os.Bundle
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.NavigationUI
-import ru.msokolov.weatherapp.databinding.FragmentMainBinding
-
-
-class MainFragment: Fragment(R.layout.fragment_main) {
-
-    private lateinit var binding: FragmentMainBinding
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding = FragmentMainBinding.bind(view)
-        (activity as AppCompatActivity?)!!.supportActionBar!!.hide()
-
-        val navHost = childFragmentManager.findFragmentById(R.id.tabsContainer) as NavHostFragment
-        val navController = navHost.navController
-        NavigationUI.setupWithNavController(binding.bottomNavigationView, navController)
-    }
-}
