--- app/src/main/res/navigation/nav_tabs.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_tabs"
-    app:startDestination="@id/now">
-
-    <include app:graph="@navigation/now" />
-    <include app:graph="@navigation/today" />
-    <include app:graph="@navigation/week" />
-    <!--<include app:graph="@navigation/settings" />-->
-</navigation>
