--- app/src/main/res/navigation/now.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/now"
-    app:startDestination="@id/nowFragment">
-
-    <fragment
-        android:id="@+id/nowFragment"
-        android:name="ru.msokolov.weatherapp.now.NowFragment"
-        android:label="fragment_now"
-        tools:layout="@layout/fragment_now" />
-</navigation>
