--- app/src/main/res/navigation/today.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/today"
-    app:startDestination="@id/todayFragment">
-
-    <fragment
-        android:id="@+id/todayFragment"
-        android:name="ru.msokolov.weatherapp.today.TodayFragment"
-        android:label="fragment_today"
-        tools:layout="@layout/fragment_today" />
-</navigation>
