--- app/src/main/res/navigation/week.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/week"
-    app:startDestination="@id/weekFragment">
-    <fragment
-        android:id="@+id/weekFragment"
-        android:name="ru.msokolov.weatherapp.week.WeekFragment"
-        android:label="WeekFragment"
-        tools:layout="@layout/fragment_week"/>
-</navigation>
