+++ feature/tomorrow/src/main/AndroidManifest.xml
