+++ feature/tomorrow/src/test/java/ru/msokolov/weatherapp/tomorrow/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package ru.msokolov.weatherapp.now
+package ru.msokolov.weatherapp.tomorrow
 
 import org.junit.Test
